MIG-14813: Moving exception and polishing up tests
