+++ phpunit.xml.dist
@@ -21,9 +21,12 @@
         </testsuite>
     </testsuites>
 
-    <source restrictDeprecations="true" restrictNotices="true" restrictWarnings="true">
+    <source restrictNotices="true" restrictWarnings="true">
         <include>
             <directory>src</directory>
         </include>
+        <exclude>
+            <directory suffix=".php">src/Arbor/Api/Exception</directory>
+        </exclude>
     </source>
 </phpunit>
