+++ src/Arbor/Api/Exception/ResourceNotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Arbor\Api\Exception;
+
+use InvalidArgumentException;
+
+class ResourceNotFoundException extends InvalidArgumentException
+{
+}
