+++ src/Arbor/Api/Gateway/PsrRestGateway.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Arbor\Api\Gateway;
 
 use Arbor\Api\Gateway\HttpClient\HttpClientInterface;
-use Arbor\Api\ServerErrorException;
+use Arbor\Api\Exception\ServerErrorException;
 use Arbor\ChangeLog\Change;
 use Arbor\Filter\CamelCaseToDash;
 use Arbor\Filter\PluralizeFilter;
