--- src/Arbor/Api/ResourceNotFoundException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Arbor\Api;
-
-class ResourceNotFoundException extends \InvalidArgumentException
-{
-}
